+++ Packages/BrandonUtils/Standalone/Optional/IFailableFunc.TValue.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+using JetBrains.Annotations;
+
+namespace BrandonUtils.Standalone.Optional {
+    /**
+     * <inheritdoc cref="IFailableFunc{TValue,TExcuse}"/>
+     */
+    [PublicAPI]
+    public interface IFailableFunc<out TValue> : IFailableFunc<TValue, Exception> { }
+}