+++ Packages/BrandonUtils/Standalone/Optional/IOptional.cs
@@ -1,6 +1,10 @@
 ﻿using System;
 
 namespace BrandonUtils.Standalone.Optional {
+    /// <summary>
+    /// A mockery of Java's <a href="https://docs.oracle.com/javase/8/docs/api/java/util/Optional.html">Optional</a> class.
+    /// </summary>
+    /// <typeparam name="T">the type of the <see cref="Value"/> that <i>might</i> be there.</typeparam>
     public interface IOptional<out T> {
         /// <summary>
         /// Whether or not a <see cref="Value"/> is present.